refactor: extracted CollectionValue into collection_value.py
